Update Batfield.py
Fixing the issue where clientside units move way faster than serverside units.

Now the server decides who dies and who wins, and it sets a seed for bullet generation. That seed can get out of sync if there are a lot of units firing at once, but the more units there are, the less the seed sync matters. It is conceivable that a game could get so out of sync that this proves a major disruption, especially in very evenly-matched high-unit games which go for a lot of rounds.... I'll tune this in later.